Add render home to the index
