+++ src/modules/helper.js
@@ -1,0 +1,9 @@
+function renderPage(page) {
+  return () => {
+    const renderArea = document.getElementById(page.renderAreaId);
+    renderArea.innerHTML = '';
+    page.render(renderArea);
+  };
+}
+
+export { renderPage };
